Initial restructuring of fasta2speclib for predict-library
