--- fasta2speclib/fasta2speclib_config.json
@@ -1,20 +1,0 @@
-{
-    "output_filetype":["msp", "mgf", "bibliospec", "dlib"],
-    "charges":[2, 3],
-    "min_peplen":8,
-    "max_peplen":30,
-    "cleavage_rule":"trypsin",
-    "missed_cleavages":2,
-    "semi_specific":false,
-    "modifications":[
-        {"name":"Acetyl", "unimod_accession":1, "mass_shift":42.010565, "amino_acid":null, "protein_n_term":true},
-        {"name":"Oxidation", "unimod_accession":35, "mass_shift":15.994915, "amino_acid":"M"},
-        {"name":"Carbamidomethyl", "unimod_accession":4, "mass_shift":57.021464, "amino_acid":"C", "fixed":true}
-    ],
-    "ms2pip_model":"HCD",
-    "decoy":false,
-    "add_retention_time":true,
-    "deeplc": {},
-    "batch_size":10000,
-    "num_cpu":null
-}
