--- fasta2speclib/fasta2speclib_config.md
@@ -1,2 +1,0 @@
-# fasta2speclib configuration info
-Moved to [compomics.github.io](http://compomics.github.io/projects/ms2pip_c/wiki/fasta2speclib.html)
